Refactor loaders to /data package
